+++ packages/editor/core/src/ui/extensions/index.tsx
@@ -141,8 +141,11 @@ export const CoreEditorExtensions = ({
     placeholder: ({ editor, node }) => {
       if (node.type.name === "heading") return `Heading ${node.attrs.level}`;
 
+      if (editor.storage.image.uploadInProgress) return "";
+
       const shouldHidePlaceholder =
         editor.isActive("table") || editor.isActive("codeBlock") || editor.isActive("image");
+
       if (shouldHidePlaceholder) return "";
 
       if (placeholder) {
