+++ packages/editor/core/src/ui/plugins/image/constants.ts
@@ -1,0 +1,7 @@
+import { PluginKey } from "@tiptap/pm/state";
+
+export const uploadKey = new PluginKey("upload-image");
+export const deleteKey = new PluginKey("delete-image");
+export const restoreKey = new PluginKey("restore-image");
+
+export const IMAGE_NODE_TYPE = "image";
