+++ packages/editor/core/src/ui/plugins/image/types/image-node.ts
@@ -1,0 +1,13 @@
+import { Node as ProseMirrorNode } from "@tiptap/pm/model";
+
+export interface ImageNode extends ProseMirrorNode {
+  attrs: {
+    src: string;
+    id: string;
+  };
+}
+
+export type ImageExtensionStorage = {
+  deletedImageSet: Map<string, boolean>;
+  uploadInProgress: boolean;
+};
